+++ viteclientts/src/components/constants.tsx
@@ -1,0 +1,42 @@
+export const daysOfWeek: string[] = ['Sonntag', 'Montag', 'Dienstag', 'Mittwoch', 'Donnerstag', 'Freitag', 'Samstag'];
+export const daysOfWeekNumbers: { [key: string]: number } = {
+  'Sonntag': 0,
+  'Montag': 1,
+  'Dienstag': 2,
+  'Mittwoch': 3,
+  'Donnerstag': 4,
+  'Freitag': 5,
+  'Samstag': 6
+};
+export const months: string[] = ['Januar', 'Februar', 'März', 'April', 'Mai', 'Juni', 'Juli', 'August', 'September', 'Oktober', 'November', 'Dezember'];
+export const monthsNumbers: { [key: string]: number } = {
+  'Januar': 0,
+  'Februar': 1,
+  'März': 2,
+  'April': 3,
+  'Mai': 4,
+  'Juni': 5,
+  'Juli': 6,
+  'August': 7,
+  'September': 8,
+  'Oktober': 9,
+  'November': 10,
+  'Dezember': 11
+};
+export const switchDescriptions: string[] = ['Stefan Nord', 'Stefan Ost', 'Lukas Süd', 'Lukas West', 'Alle'];
+export const bewaesserungsTopics: string[] = [
+  'bewaesserung/switch/stefanNord',
+  'bewaesserung/switch/stefanOst',
+  'bewaesserung/switch/lukasSued',
+  'bewaesserung/switch/lukasWest',
+  'bewaesserung/switch/alle',
+];
+export const zoneOrder: string[] = ["Stefan Nord", "Stefan Ost", "Lukas Süd", "Lukas West"];
+
+export const bewaesserungsTopicsSet: string[] = [
+  'bewaesserung/switch/stefanNord/set',
+  'bewaesserung/switch/stefanOst/set',
+  'bewaesserung/switch/lukasSued/set',
+  'bewaesserung/switch/lukasWest/set',
+  'bewaesserung/switch/alle/set',
+];